Added event on migration
This new event allows plugins to act when a database migration is
executed. This makes it possible to do more complex migrations that
require more than just SQL.

The BEFORE event is executed before the migration file is applied. The
default action is to apply that file and increase the dbversion field in
the opts table. When an event prevents the default, the proper result
has to be set! When the result is true, the dbversion is increased and
any additional migrations are run. When the result is false, the
migration process is aborted.

The AFTER event is only triggered after the migration file has been
applied or the BEFORE event set a truish result. At the AFTER time, the
dbversion has already been increased.