Fix import/export handling, always split multi-statements
Even though mutliple statements can be passed to $sth->execute(),
failures will only be checked in the very first statement. Failures
later on will lead to the rest of the statements to be silently ignored
rendering our rollback strategy moot.

We now always split multiline statements and run them separately
(affects import and migrations).

The export was adjusted to not include the transaction handling and
fix entry exports.